+++ project/src/components/details-tab.tsx
@@ -1,0 +1,39 @@
+import {Film} from '../types/film-type';
+
+type Props={
+  film: Film;
+}
+
+function DetailsTab ({film}: Props) : JSX.Element {
+  return(
+    <div className="film-card__text film-card__row">
+      <div className="film-card__text-col">
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Director</strong>
+          <span className="film-card__details-value">{film.director}</span>
+        </p>
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Starring</strong>
+          <span className="film-card__details-value">
+            {film.starring.join(', ')}
+          </span>
+        </p>
+      </div>
+      <div className="film-card__text-col">
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Run Time</strong>
+          <span className="film-card__details-value">{film.runTime}</span>
+        </p>
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Genre</strong>
+          <span className="film-card__details-value">{film.genre}</span>
+        </p>
+        <p className="film-card__details-item">
+          <strong className="film-card__details-name">Released</strong>
+          <span className="film-card__details-value">{film.released}</span>
+        </p>
+      </div>
+    </div>
+  );
+}
+export default DetailsTab;
