+++ project/src/components/video-player.tsx
