+++ project/src/pages/my-list-page.tsx
@@ -1,5 +1,5 @@
 import {Film} from '../types/film-type';
-import FilmsListComponent from '../components/film-list-component';
+import FilmsListComponent from '../components/film-list';
 
 type Props = {
   films: Film[];
